Throw known error if missing mediaType in __eq__
Even if it's contrary to what the spec says, a very small number of
v2 images do not have mediaType defined. We could go smart and try to
guess the type inspecting the manifest and look for either "layers"
 or "manifests" but given the low amount of images that have this
behaviour (in app-interface we found 20 of 16768 image tags synced,
belonging to just two repos) and the fact that this is contrary to
the spec, it looks better to throw an ImageComparisonError Exception.

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>