fix stuck threadpool if thread terminates with sys.exit
if a thread managed by `threading.run` with `return_exceptions=True` terminates with `sys.exit`, the unterlying threadpool will be stuck and `run` will never return.
the reason is the error handling in `_catching_traceback`, which catches only `Exception` classes while `sys.exit` yields a `SystemExit` exception that does not inherit from `Exception`.

the solving approach catches `SystemExit` explicitely and interprets `SystemExit.code`. if the code indicates an issue (everything not 0 and not `False`), the exception is returned. any code that is 0 or `False` indicates success without a return value, so `None` is returned.

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>